requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
openai==1.59.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.5
aiofiles==24.1.0
python-dotenv==1.0.1
